tried to implement background audio but its not working
i tried to implement the background text to speech and got told it is not supported. tried to implement the playing of mp3 files, and that did not work while locked. this effor is not worth it and would rather use a new issue to implement displaying of the name on the apple watch